Move login/logout into new helper
This code is about to be reused in the subscription authentication
controller, so it makes sense to move it out now.
